+++ docs/source/news.rst
@@ -10,6 +10,7 @@ What's New
   Contributed by Jonathan Jacobs.
 * ``write_traceback`` and ``writeFailure`` no longer require a ``system`` argument, as the combination of traceback and action context should suffice to discover the origin of the problem.
   This is a minor change to output format as the field is also omitted from the resulting ``eliot:traceback`` messages.
+* The ``validate_logging`` testing utility now skips validation when the decorated test method raises ``SkipTest``.
 
 0.6.0
 ^^^^^